backend/requirements.txt
Django==5.1
django-ninja==1.3.0
python-dotenv==1.0.1
uvicorn==0.30.6
psycopg2-binary==2.9.9